refactor: Add omnisearch to tissue filter (#2921)
