+++ frontend/src/components/common/Filter/components/FilterContent/components/FilterMenu/style.ts
@@ -2,7 +2,7 @@ import { Classes, MenuItem } from "@blueprintjs/core";
 import { Divider } from "@material-ui/core";
 import { GRAY, PT_TEXT_COLOR } from "src/components/common/theme";
 import styled from "styled-components";
-import { scrollbar } from "../../common/style";
+import { scrollbar } from "../../../../common/style";
 
 export const MAX_DISPLAYABLE_MENU_ITEMS = 9;
 const DIVIDER_HEIGHT_PX = 9;
