+++ frontend/src/components/common/Filter/components/FilterContent/components/FilterRange/style.ts
