+++ frontend/src/components/common/Filter/components/FilterContent/components/FilterViews/components/FilterMultiPanelCategoryView/style.ts
