+++ frontend/src/components/common/Filter/components/FilterContent/components/FilterViews/components/FilterView/index.tsx
@@ -4,16 +4,16 @@ import { useResizeObserver } from "src/common/hooks/useResizeObserver";
 import {
   OnFilterFn,
   OntologyCategoryTreeNodeView,
-  OnUpdateSearchValueFn,
 } from "src/components/common/Filter/common/entities";
 import {
   ViewDivider,
   ViewHeader,
   ViewPanel,
   ViewPanelScroll,
-} from "src/components/common/Filter/components/FilterViews/components/FilterView/style";
-import FilterViewList from "src/components/common/Filter/components/FilterViews/components/FilterViewList";
-import FilterViewSearch from "src/components/common/Filter/components/FilterViews/components/FilterViewSearch";
+} from "src/components/common/Filter/components/FilterContent/components/FilterViews/components/FilterView/style";
+import FilterViewList from "src/components/common/Filter/components/FilterContent/components/FilterViews/components/FilterViewList";
+import FilterSearch from "src/components/common/Filter/components/FilterSearch";
+import { useFilterSearch } from "src/components/common/Filter/components/FilterSearch/common/useFilterSearch";
 
 const ADDITIONAL_PANEL_WIDTH = 8;
 
@@ -28,7 +28,6 @@ interface Props {
   isSearchable: boolean;
   isZerosVisible: boolean;
   onFilter: OnFilterFn;
-  onUpdateSearchValue: OnUpdateSearchValueFn;
   showViewDivider: boolean;
   values: OntologyCategoryTreeNodeView[];
   viewListMaxHeight: number;
@@ -40,17 +39,16 @@ export default function FilterView({
   isSearchable,
   isZerosVisible,
   onFilter,
-  onUpdateSearchValue,
   showViewDivider,
   values,
   viewListMaxHeight,
 }: Props): JSX.Element {
+  const { searchValue, setSearchValue } = useFilterSearch();
   const panelRef = useRef<HTMLDivElement>(null);
   const listContainerRef = useRef<HTMLDivElement>(null);
   const listContainerRect = useResizeObserver(listContainerRef);
   const [panelScrollable, setPanelScrollable] = useState(false);
   const [panelWidth, setPanelWidth] = useState<number>(0);
-  const [searchValue, setSearchValue] = useState<string>("");
   const { scrollHeight: listScrollHeight } = listContainerRect || {};
   const filteredValues = filterViewValues(values, searchValue);
 
@@ -75,8 +73,8 @@ export default function FilterView({
       <ViewPanel panelWidth={panelWidth} ref={panelRef}>
         {/* Optional search bar */}
         {isSearchable && (
-          <FilterViewSearch
-            onUpdateSearchValue={onUpdateSearchValue}
+          <FilterSearch
+            searchValue={searchValue}
             setSearchValue={setSearchValue}
           />
         )}
