+++ frontend/src/components/common/Filter/components/FilterContent/components/FilterViews/components/FilterView/style.ts
