+++ frontend/src/components/common/Filter/components/FilterContent/components/FilterViews/components/FilterViewList/style.ts
