+++ frontend/src/components/common/Filter/components/FilterContent/components/FilterViews/index.tsx
@@ -1,17 +1,17 @@
+import React from "react";
 import { CategoryFilterId } from "src/common/hooks/useCategoryFilter";
 import {
   OnFilterFn,
   OntologyCategoryTreeView,
-  OnUpdateSearchValueFn,
 } from "src/components/common/Filter/common/entities";
 import FilterView, {
   MAX_DISPLAYABLE_LIST_ITEMS,
-} from "src/components/common/Filter/components/FilterViews/components/FilterView";
+} from "src/components/common/Filter/components/FilterContent/components/FilterViews/components/FilterView";
 import {
   VIEW_LIST_ITEM_HEIGHT,
   VIEW_LIST_SUBHEADER_HEIGHT,
-} from "src/components/common/Filter/components/FilterViews/components/FilterView/style";
-import { ViewsMenu } from "src/components/common/Filter/components/FilterViews/style";
+} from "src/components/common/Filter/components/FilterContent/components/FilterViews/components/FilterView/style";
+import { ViewsMenu } from "src/components/common/Filter/components/FilterContent/components/FilterViews/style";
 
 export const enum CATEGORY_VIEWS_QUANTIFIER {
   NON_SINGLETON = "NON_SINGLETON",
@@ -23,7 +23,6 @@ interface Props {
   isSearchable: boolean;
   isZerosVisible: boolean;
   onFilter: OnFilterFn;
-  onUpdateSearchValue: OnUpdateSearchValueFn;
   views: OntologyCategoryTreeView[];
 }
 
@@ -32,7 +31,6 @@ export default function FilterViews({
   isSearchable,
   isZerosVisible,
   onFilter,
-  onUpdateSearchValue,
   views,
 }: Props): JSX.Element {
   const viewsToDisplay = views.filter(
@@ -56,7 +54,6 @@ export default function FilterViews({
           isSearchable={isSearchable}
           isZerosVisible={isZerosVisible}
           onFilter={onFilter}
-          onUpdateSearchValue={onUpdateSearchValue}
           showViewDivider={i !== 0}
           values={children}
           viewListMaxHeight={viewListMaxHeight}
