+++ frontend/src/components/common/Filter/index.tsx
@@ -1,26 +1,16 @@
-import { ChangeEvent, ReactNode } from "react";
 import { CategoryFilterId } from "src/common/hooks/useCategoryFilter";
 import {
-  CategoryValueKey,
   CategoryView,
   OnFilterFn,
   OntologyCategoryView,
   OntologyMultiPanelCategoryView,
-  Range,
   RangeCategoryView,
-  SelectCategoryValueView,
   SelectCategoryView,
-  SetSearchValueFn,
 } from "src/components/common/Filter/common/entities";
 import { formatNumberToScale } from "src/components/common/Filter/common/utils";
 import BasicFilter from "src/components/common/Filter/components/BasicFilter";
 import FilterLabel from "src/components/common/Filter/components/FilterLabel";
-import FilterMenu from "src/components/common/Filter/components/FilterMenu";
-import { MAX_DISPLAYABLE_MENU_ITEMS } from "src/components/common/Filter/components/FilterMenu/style";
-import FilterRange from "src/components/common/Filter/components/FilterRange";
-import { FilterSearchState } from "src/components/common/Filter/components/FilterSearch/common/useFilterSearch";
-import FilterViews from "src/components/common/Filter/components/FilterViews";
-import FilterMultiPanelCategoryView from "src/components/common/Filter/components/FilterViews/components/FilterMultiPanelCategoryView";
+import FilterContent from "./components/FilterContent";
 import FilterTags, { CategoryTag } from "./components/FilterTags";
 
 interface Props {
@@ -38,110 +28,24 @@ export default function Filter({
         const { isDisabled = false, label, tooltip } = categoryView;
         return (
           <BasicFilter
-            Content={(filterSearchState: FilterSearchState): JSX.Element =>
-              buildBasicFilterContent(categoryView, onFilter, filterSearchState)
+            content={
+              <FilterContent categoryView={categoryView} onFilter={onFilter} />
             }
-            flipEnabled={categoryViews.label !== "Tissue (Ontology)"} // TODO(cc) review use of flipEnabled prop
+            flipEnabled={categoryView.label !== "Tissue (Ontology)"} // TODO(cc) review use of flipEnabled prop
             isDisabled={isDisabled}
             key={categoryView.label}
             tags={<FilterTags tags={buildFilterTags(categoryView, onFilter)} />}
-            target={buildFilterLabel(label, isDisabled, tooltip)}
+            target={
+              <FilterLabel
+                isDisabled={isDisabled}
+                label={label}
+                tooltip={tooltip}
+              />
+            }
           />
         );
       })}
     </>
-  );
-}
-
-/**
- * Build content model of basic filter depending on category type.
- * @param categoryView - View model of category to display.
- * @param onFilter - Function to execute on select of category value or remove of selected category value.
- * @param filterSearchState - Filter search value and corresponding functions to clear or update search value.
- * @returns Element representing content to display inside basic filter menu.
- */
-function buildBasicFilterContent(
-  categoryView: CategoryView,
-  onFilter: OnFilterFn,
-  filterSearchState: FilterSearchState
-): JSX.Element {
-  const { key } = categoryView;
-  const { clearSearchValueFn, searchValue, setSearchValue } = filterSearchState;
-
-  // Update onFilter function with clear search value function.
-  const onFilterWithClearSearch = (
-    categoryKey: CategoryFilterId,
-    key: CategoryValueKey | null, // null for ranges.
-    values: CategoryValueKey[] | Range
-  ) => {
-    onFilter(categoryKey, key, values);
-    clearSearchValueFn();
-  };
-
-  // Handle ontology categories.
-  if (isOntologyCategoryView(categoryView)) {
-    return (
-      <FilterViews
-        categoryKey={key}
-        isSearchable={categoryView.isSearchable}
-        isZerosVisible={categoryView.isZerosVisible}
-        onFilter={onFilter}
-        onUpdateSearchValue={onUpdateSearchValue}
-        views={categoryView.views}
-      />
-    );
-  }
-
-  // Handle select categories
-  if (isSelectCategoryView(categoryView)) {
-    const { pinnedValues, unpinnedValues, values } = categoryView;
-    return (
-      <FilterMenu
-        categoryKey={key}
-        isMultiselect // Can possibly be single select with future filter types
-        isSearchable={values.length > MAX_DISPLAYABLE_MENU_ITEMS}
-        onFilter={onFilter}
-        onUpdateSearchValue={onUpdateSearchValue}
-        pinnedValues={filterCategoryValuesWithCount(pinnedValues)}
-        searchValue={searchValue}
-        setSearchValue={setSearchValue}
-        unpinnedValues={filterCategoryValuesWithCount(unpinnedValues)}
-        values={filterCategoryValuesWithCount(values)}
-      />
-    );
-  }
-
-  // Handle ontology multi panel categories
-  if (isOntologyMultiPanelCategoryView(categoryView)) {
-    return (
-      <FilterMultiPanelCategoryView
-        categoryView={categoryView}
-        isSearchable={isFilterMultiPanelSearchable(categoryView)}
-        onFilter={onFilterWithClearSearch}
-        searchValue={searchValue}
-        setSearchValue={setSearchValue}
-      />
-    );
-  }
-
-  // Otherwise, handle range categories
-  return <FilterRange categoryView={categoryView} onFilter={onFilter} />;
-}
-
-/**
- * Build the filter label for the given category.
- * @param label - Category view label.
- * @param isDisabled - True if this category view is currently disabled.
- * @param tooltip - Category tooltip.
- * @returns React node representing content to display as filter label.
- */
-function buildFilterLabel(
-  label: string,
-  isDisabled: boolean,
-  tooltip?: string
-): ReactNode {
-  return (
-    <FilterLabel isDisabled={isDisabled} label={label} tooltip={tooltip} />
   );
 }
 
@@ -292,36 +196,6 @@ function createRangeTagLabel(min: number, max: number): [string, string] {
 }
 
 /**
- * Returns filtered category values where category count is greater than zero.
- * @param categoryValues - Category value view models for a given category.
- * @returns category values with a count
- */
-function filterCategoryValuesWithCount(
-  categoryValues: SelectCategoryValueView[]
-): SelectCategoryValueView[] {
-  return categoryValues.filter(({ count }) => count > 0);
-}
-
-/**
- * Returns true if ontology multi panel is searchable.
- * @param categoryView - Ontology multi panel category view.
- * @returns true if ontology multi panel is searchable.
- */
-function isFilterMultiPanelSearchable(
-  categoryView: OntologyMultiPanelCategoryView
-): boolean {
-  let isSearchable = false;
-  for (const panel of categoryView.panels) {
-    const { views } = panel;
-    if (views.length > MAX_DISPLAYABLE_MENU_ITEMS) {
-      isSearchable = true;
-      break;
-    }
-  }
-  return isSearchable;
-}
-
-/**
  * Determine if the given category view is an ontology category view and not a select or range or ontology multi panel category view.
  * @param categoryView - Selected filter value, either a category value key (e.g. "normal"), range (e.g. [0, 10]) or
  * ontology tree.
@@ -344,19 +218,6 @@ export function isOntologyMultiPanelCategoryView(
 }
 
 /**
- * Determine if the given category view is a range category view and not a select or ontology with or without multi panel category view.
- * TODO(cc) unused function.
- * @param categoryView - Selected filter value, either a category value key (e.g. "normal"), range (e.g. [0, 10]) or
- * ontology tree.
- * @returns True if the given category view is a range category view.
- */
-export function isRangeCategoryView(
-  categoryView: CategoryView
-): categoryView is RangeCategoryView {
-  return (categoryView as RangeCategoryView).max !== undefined;
-}
-
-/**
  * Determine if the given category view is a selected category view and not an ontology with or without multi panel or range category view.
  * @param categoryView - Selected filter value, either a category value key (e.g. "normal"), range (e.g. [0, 10]) or
  * ontology tree.
@@ -366,16 +227,4 @@ export function isSelectCategoryView(
   categoryView: CategoryView
 ): categoryView is SelectCategoryView {
   return (categoryView as SelectCategoryView).values !== undefined;
-}
-
-/**
- * Sets state searchValue with updated search value.
- * @param changeEvent
- * @param setSearchValue
- */
-function onUpdateSearchValue(
-  changeEvent: ChangeEvent<HTMLInputElement>,
-  setSearchValue: SetSearchValueFn
-): void {
-  setSearchValue(changeEvent.target.value.toLowerCase());
 }
